+++ y60/sdlengine/SDLTextRenderer.h
@@ -132,7 +132,7 @@ namespace y60 {
                 bool          newline;
                 double        kerning;
                 int           minx;
-                std::vector<float> sdl_x_position;
+                std::vector<int> sdl_x_position;
                 private:
                 Word() {}
             };
